#156 [feat] : 퀘스트 화면 내에서 모든 퀘스트 완료 시 오프보딩 모달 노출 기능 추가

+++ app/src/main/java/com/byeboo/app/domain/model/quest/QuestStateModel.kt
@@ -3,7 +3,7 @@ package com.byeboo.app.domain.model.quest
 data class QuestStateModel(
     val todayComplete: Boolean,
     val userCurrentStatus: String,
-    val count: Int
+    val count: Long
 )
 
 data class QuestDialogue(
